refactor: architecture
