+++ src/db/userdb.ts
@@ -1,0 +1,8 @@
+export type User = {
+  id?: string;
+  username: string;
+  age: number;
+  hobbies: string[];
+};
+
+export const users: User[] = [];
