+++ src/routes/types.ts
@@ -1,0 +1,9 @@
+import { Handler } from "../utils/getHandlerAndParam";
+
+export type Route = {
+  [key: string]: Handler;
+};
+
+export type Routes = {
+  [key: string]: Route;
+};
